src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements/base.in requirements/dev.in requirements/release.in
#
astroid==1.5.3            # via pylint, pylint-plugin-utils
behave-django==0.5.0
behave==1.2.5             # via behave-django
blessings==1.6            # via curtsies
bpython==0.16
brotlipy==0.7.0
certifi==2017.7.27.1      # via requests
cffi==1.11.2              # via brotlipy
chardet==3.0.4            # via requests
click==6.7                # via pip-tools, safety
curtsies==0.2.11          # via bpython
decorator==4.1.2          # via ipython, traitlets
dj-database-url==0.4.2
django-debug-toolbar==1.8
django-extensions==1.9.6
django-filter==1.0.4
django-jsonfield==1.0.1
django-money==0.11.4
django-pdb==0.6.1
django-settings-export==1.2.1
django==1.11.6
djangorestframework-bulk==0.2.1
djangorestframework==3.6.4
dparse==0.2.1             # via safety
factory-boy==2.9.2
faker==0.8.6              # via factory-boy
first==2.0.1              # via pip-tools
flake8==3.4.1
greenlet==0.4.12          # via bpython
gunicorn==19.7.1
idna==2.6                 # via requests
ipython-genutils==0.2.0   # via traitlets
ipython==6.2.1
isort==4.2.15             # via pylint
jedi==0.11.0              # via ipython
lazy-object-proxy==1.3.1  # via astroid
lizard==1.12.15
mccabe==0.6.1             # via flake8, pylint
mypy==0.530
packaging==16.8           # via dparse, safety
parse-type==0.4.2         # via behave
parse==1.8.2              # via behave, parse-type
parso==0.1.0              # via jedi
pexpect==4.2.1            # via ipython
pickleshare==0.7.4        # via ipython
pip-tools==1.10.1
prompt-toolkit==1.0.15    # via ipython
psycopg2==2.6.2
ptyprocess==0.5.2         # via pexpect
pudb==2017.1.4
py-moneyed==0.7.0         # via django-money
pycodestyle==2.3.1        # via flake8
pycparser==2.18           # via cffi
pyflakes==1.5.0
pygments==2.2.0           # via bpython, ipython, pudb
pylint-django==0.7.2
pylint-plugin-utils==0.2.6  # via pylint-django
pylint==1.7.4
pyparsing==2.2.0          # via packaging
python-dateutil==2.6.1    # via faker
pytz==2017.2              # via django
pyyaml==3.12              # via dparse
requests==2.18.4          # via bpython, safety
safety==1.5.1
selenium==3.6.0
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via astroid, behave, bpython, django-extensions, dparse, faker, packaging, parse-type, pip-tools, prompt-toolkit, pylint, python-dateutil, traitlets
sqlparse==0.2.4           # via django-debug-toolbar
text-unidecode==1.0       # via faker
traitlets==4.3.2          # via ipython
typed-ast==1.1.0          # via mypy
urllib3==1.22             # via requests
urwid==1.3.1              # via pudb
wcwidth==0.1.7            # via curtsies, prompt-toolkit
werkzeug==0.12.2
whitenoise==3.3.1
wrapt==1.10.11            # via astroid
